backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
email-validator==2.1.0
requests==2.31.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
PyJWT==2.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0